fix: Update import statements and remove unnecessary code
The import statements for the `Link` and `Pressable` components from `expo-router` have been removed as they are no longer used in the code.

The `title` prop for the `Tabs.Screen` component has been updated from "Tab One" to "Home".

The headerRight prop for the `Tabs.Screen` component has been removed as it is no longer used in the code.

The `useQuery` hook and other related code have been removed from the `ModalScreen` component.

The `Details` component has been added, which includes a new query for a single book and displays the book details.

The `cleanDescription` function has been imported from the `~/utils/cleanDescription` module.

The `authors` property in the `Book` schema has been updated to default to an array with a single value of "Unknown".

The `description` property in the `Book` schema has been updated to transform the value